Sounds stuff initial upload
This is just the assets folder - if memory serves, you should be able to drop that into your project and use it, but if not, let me know. You may have to assign the sounds in the Unity editor but that's as easy as clicking and dragging. Call the PlaySounds method with the name of the sound you want to play. I currently have it set so qwertyuiopa plays sounds just for testing purposes.